requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
PyJWT==2.9.0
pydantic==2.9.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
boto3==1.35.10
